Solved using_as.rs

+++ conversions/using_as.rs
@@ -5,11 +5,12 @@
 // The goal is to make sure that the division does not fail to compile
 // and returns the proper type.
 
-// I AM NOT DONE
+// // I AM NOT DONE
 
 fn average(values: &[f64]) -> f64 {
     let total = values.iter().fold(0.0, |a, b| a + b);
-    total / values.len()
+    // SOLUTION: Added `as f64` in following line
+    total / (values.len() as f64)
 }
 
 fn main() {
